+++ src/json/cJSON.c
@@ -318,7 +318,7 @@ loop_end:
 
     if ( !is_float )
     {
-        if ( number >= LLONG_MAX || number <= LLONG_MIN )
+        if ( number >= (double)(LLONG_MAX) || number <= LLONG_MIN )
 		    is_float = 1;
         else
         {
@@ -335,7 +335,7 @@ loop_end:
 	if ( is_float )
 	{
         /* use saturation in case of overflow */
-        if (number >= LLONG_MAX)
+        if (number >= (double)(LLONG_MAX))
         {
             item->valueint = LLONG_MAX;
         }
@@ -363,7 +363,7 @@ loop_end:
 /* don't ask me, but the original cJSON_SetNumberValue returns an integer or double */
 CJSON_PUBLIC(double) cJSON_SetNumberHelper(cJSON *object, double number)
 {
-    if (number >= LLONG_MAX)
+    if (number >= (double)LLONG_MAX)
     {
         object->valueint = LLONG_MAX;
     }
@@ -2231,7 +2231,7 @@ CJSON_PUBLIC(cJSON *) cJSON_CreateNumber(double num)
         item->valuedouble = num;
 
         /* use saturation in case of overflow */
-        if (num >= LLONG_MAX)
+        if (num >= (double)LLONG_MAX)
         {
             item->valueint = LLONG_MAX;
         }
